backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.0
requests>=2.31.0
httpx>=0.25.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.0.0
